move collectFiles to shared directory
