+++ packages/core/src/amazonqFeatureDev/session/sessionState.ts
@@ -26,7 +26,7 @@ import {
     SessionStateInteraction,
     SessionStatePhase,
 } from '../types'
-import { collectFiles, getWorkspaceFoldersByPrefixes, prepareRepoData } from '../util/files'
+import { prepareRepoData } from '../util/files'
 import { TelemetryHelper } from '../util/telemetryHelper'
 import { uploadCode } from '../util/upload'
 import { CodeReference } from '../../amazonq/webview/ui/connector'
@@ -34,6 +34,7 @@ import { isPresent } from '../../shared/utilities/collectionUtils'
 import { encodeHTML } from '../../shared/utilities/textUtilities'
 import { AuthUtil } from '../../codewhisperer/util/authUtil'
 import { randomUUID } from '../../common/crypto'
+import { collectFiles, getWorkspaceFoldersByPrefixes } from '../../shared/utilities/workspaceUtils'
 
 export class ConversationNotStartedState implements Omit<SessionState, 'uploadId'> {
     public tokenSource: vscode.CancellationTokenSource
