+++ packages/core/src/shared/utilities/gitignore.ts
